modularized, 21second run
