+++ lib/sensors.h
@@ -1,0 +1,4 @@
+ int get_dist()
+  {
+    return (dhigh - analogRead(AD_PIN))*dscale;
+  }
